Expressions_Tamika Mosby
I was having a hard time with usderstanding why I wasn't getting the
same console.log output as the video. Could you please explain to me
what I was doing wrong. I was trying to chec the red box on the right
to see what error I was getting. This is also commented in the file.